Refactor to more appropriate location
